Foi criado o modulo profile para organizar as features de profile do usuario
