+++ front/templates/users/imagem_profile.html
@@ -16,7 +16,7 @@
                 </div>
                 <!-- /.card-header -->
                 <!-- form start -->
-                <form role="form" action="{{ url_for('update_profile_image' )}}" method="post" enctype="multipart/form-data">
+                <form role="form" action="{{ url_for('profile.update_profile_image' )}}" method="post" enctype="multipart/form-data">
                     <div class="card-body">
                         <div class="form-group">
                             <label for="Picture">Picture</label><br>
